feat: adaptar backend para deploy na Vercel com serverless-http
